+++ click_extra/tests/test_config.py
@@ -38,6 +38,8 @@ from click_extra.decorators import config_option, extra_group
 from .conftest import (
     default_debug_uncolored_log_end,
     default_debug_uncolored_log_start,
+    default_debug_uncolored_logging,
+    default_debug_uncolored_version_details,
 )
 
 DUMMY_TOML_FILE, DUMMY_TOML_DATA = (
@@ -463,11 +465,9 @@ def test_conf_file_overrides_defaults(
                 rf'"GET /{re.escape(conf_name)} HTTP/1\.1" 200 -\n'
             )
         debug_log += (
-            r"debug: Set <Logger click_extra \(DEBUG\)> to DEBUG.\n"
-            r"debug: Set <RootLogger root \(DEBUG\)> to DEBUG.\n"
-            r"debug: \S+, version \S+\n"
-            r"debug: {.*}\n"
-            rf"{default_debug_uncolored_log_end}"
+            default_debug_uncolored_logging +
+            default_debug_uncolored_version_details +
+            default_debug_uncolored_log_end
         )
         assert re.fullmatch(debug_log, result.stderr)