+++ click_extra/tests/test_timer.py
@@ -14,7 +14,8 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 """Test defaults of our custom commands, as well as their customizations and attached
-options, and how they interact with each others."""
+options, and how they interact with each others.
+"""
 
 from __future__ import annotations